requirements.txt
python-dxf
tqdm
